requirements.txt
# API Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Traitement d'images
Pillow==10.2.0
opencv-python==4.9.0.80

# Calcul scientifique
numpy==1.26.3

# Utilitaires
python-dotenv==1.0.0

# Notebooks (optionnel)
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.1
pandas==2.1.4
scikit-learn==1.4.0